+++ src/exponential.cpp
@@ -1,5 +1,6 @@
 //@author Erik Edwards
-//@date 2019-2020
+//@date 2018-present
+//@license BSD 3-clause
 
 
 #include <iostream>
@@ -10,7 +11,7 @@
 #include <valarray>
 #include <unordered_map>
 #include <argtable2.h>
-#include "/home/erik/codee/util/cmli.hpp"
+#include "../util/cmli.hpp"
 #include <random>
 #include <complex>
 
@@ -29,8 +30,8 @@ int main(int argc, char *argv[])
     const string errstr = ": \033[1;31merror:\033[0m ";
     const string warstr = ": \033[1;35mwarning:\033[0m ";
     const string progstr(__FILE__,string(__FILE__).find_last_of("/")+1,strlen(__FILE__)-string(__FILE__).find_last_of("/")-5);
-    const valarray<uint8_t> oktypes = {1,2,101,102};
-    const size_t O = 1;
+    const valarray<size_t> oktypes = {1u,2u,101u,102u};
+    const size_t O = 1u;
     ofstream ofs1;
     int8_t stdo1, wo1;
     ioinfo o1;
@@ -86,15 +87,15 @@ int main(int argc, char *argv[])
     //Get options
 
     //Get o1.F
-    if (a_ofmt->count==0) { o1.F = 147; }
+    if (a_ofmt->count==0) { o1.F = 147u; }
     else if (a_ofmt->ival[0]<0) { cerr << progstr+": " << __LINE__ << errstr << "output file format must be nonnegative" << endl; return 1; }
     else if (a_ofmt->ival[0]>255) { cerr << progstr+": " << __LINE__ << errstr << "output file format must be < 256" << endl; return 1; }
-    else { o1.F = uint8_t(a_ofmt->ival[0]); }
+    else { o1.F = size_t(a_ofmt->ival[0]); }
 
     //Get o1.T
-    if (a_otyp->count==0) { o1.T = 1; }
+    if (a_otyp->count==0) { o1.T = 1u; }
     else if (a_otyp->ival[0]<1) { cerr << progstr+": " << __LINE__ << errstr << "output data type must be positive" << endl; return 1; }
-    else { o1.T = uint8_t(a_otyp->ival[0]); }
+    else { o1.T = size_t(a_otyp->ival[0]); }
     if ((o1.T==oktypes).sum()==0)
     {
         cerr << progstr+": " << __LINE__ << errstr << "input data type must be in " << "{";
@@ -145,7 +146,7 @@ int main(int argc, char *argv[])
 
 
     //Process
-    if (o1.T==1)
+    if (o1.T==1u)
     {
         valarray<float> Y(o1.N());
         exponential_distribution<float> distr(lam);
@@ -171,7 +172,7 @@ int main(int argc, char *argv[])
             catch (...) { cerr << progstr+": " << __LINE__ << errstr << "problem writing output file (Y)" << endl; return 1; }
         }
     }
-    else if (o1.T==101)
+    else if (o1.T==101u)
     {
         valarray<complex<float>> Y(o1.N());
         exponential_distribution<float> distr(lam);
@@ -184,7 +185,7 @@ int main(int argc, char *argv[])
             catch (...) { cerr << progstr+": " << __LINE__ << errstr << "problem writing output file (Y)" << endl; return 1; }
         }
     }
-    else if (o1.T==102)
+    else if (o1.T==102u)
     {
         valarray<complex<double>> Y(o1.N());
         exponential_distribution<double> distr(lam);
