+++ srci/adiff.cpp
@@ -2,8 +2,8 @@
 #include "adiff.c"
 
 //Declarations
-const valarray<uint8_t> oktypes = {1,2,101,102};
-const size_t I = 2, O = 1;
+const valarray<size_t> oktypes = {1u,2u,101u,102u};
+const size_t I = 2u, O = 1u;
 
 //Description
 string descr;
@@ -35,7 +35,7 @@ if (!major_compat(i1,i2)) { cerr << progstr+": " << __LINE__ << errstr << "input
 
 //Set output header info
 o1.F = i1.F;
-o1.T = i1.isreal() ? i1.T : i1.T-100;
+o1.T = i1.isreal() ? i1.T : i1.T-100u;
 o1.R = (i1.R>i2.R) ? i1.R : i2.R;
 o1.C = (i1.C>i2.C) ? i1.C : i2.C;
 o1.S = (i1.S>i2.S) ? i1.S : i2.S;
@@ -44,7 +44,7 @@ o1.H = (i1.H>i2.H) ? i1.H : i2.H;
 //Other prep
 
 //Process
-if (i1.T==1)
+if (i1.T==1u)
 {
     float *X1, *X2;
     try { X1 = new float[i1.N()]; }
@@ -81,7 +81,7 @@ if (i1.T==1)
     }
     delete[] X1; delete[] X2;
 }
-else if (i1.T==101)
+else if (i1.T==101u)
 {
     float *X1, *X2, *Y;
     try { X1 = new float[2u*i1.N()]; }
